+++ dje/registration.py
@@ -126,9 +126,8 @@ class DejaCodeRegistrationForm(RegistrationFormUniqueEmail):
 
         self.fields["hcaptcha"].label = ""
 
-        self.fields["updates_email_notification"].label = (
-            "Receive updates on DejaCode features and news"
-        )
+        notification_label = "Receive updates on DejaCode features and news"
+        self.fields["updates_email_notification"].label = notification_label
 
         for field in self.fields.values():
             field.help_text = None